% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduct_experiment.R
\name{compare_vectors}
\alias{compare_vectors}
\title{compare two SNP vectors}
\usage{
compare_vectors(genome_vector, pileup_vector, equivalence_mat)
}
\arguments{
\item{genome_vector}{a Snp vector of an individual genome}

\item{pileup_vector}{an Rnaseq Snp vector}

\item{equivalence_mat}{the scoring matrix which defines what is different
from what. Eg, Is A/T == T/A (should be). What about A and A/T?}
}
\value{
fraction of matches over total number of genotyped bases in both
vectors
}
\description{
compare two SNP vectors
}
