% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduct_experiment.R
\name{conduct_experiment}
\alias{conduct_experiment}
\title{Conduct an experiment to attempt to map RNAseq librarys to genomes}
\usage{
conduct_experiment(db_path, pileup_database_dir, sample, filename_map)
}
\arguments{
\item{db_path}{path to the sample_genotype_db}

\item{pileup_database_dir}{directory which stores the mislabelled and control
pileup databases}

\item{sample}{the name of the sample (a view in the sample_geontype database)}

\item{filename_map}{a dataframe which maps pileup database filenames to
id and subject numbers}
}
\value{
the results of the experiment
}
\description{
Conduct an experiment to attempt to map RNAseq librarys to genomes
}
