% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_high_expr_granges.R
\name{create_plink_region_filter}
\alias{create_plink_region_filter}
\title{Create a plink region filter from GRangeList}
\usage{
create_plink_region_filter(grl)
}
\arguments{
\item{grl}{a GRangeList object. must have type with at least factor level
gene and seqnames, start, end and gene_id}
}
\value{
a dataframe which may be written out to act as a plink filter
}
\description{
Create a plink region filter from GRangeList
}
