% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_genotype_db_functions.R
\name{create_sample_geno_db}
\alias{create_sample_geno_db}
\title{Create a sqlite database to store sample/genotype data}
\usage{
create_sample_geno_db(db_path)
}
\arguments{
\item{db_path}{a path, eg /path/to/sample_geno_db.sqlite, to where you'd like
to create the sqlite database file}
}
\description{
Create a sqlite database to store sample/genotype data
}
