% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduct_experiment.R
\name{equivalence_class_dist_matrix}
\alias{equivalence_class_dist_matrix}
\title{Create variants of the distance matrix}
\usage{
equivalence_class_dist_matrix(lenient = FALSE)
}
\arguments{
\item{lenient}{Set to TRUE to consider, for example,
A == A, A/T, T/A, A/G, G/A, A/C, C/A for a distance score of 0. Default is
FALSE}
}
\value{
a symmetric matrix with rownames corresponding to the 16 possible
genotypes
}
\description{
distance between two SNPS vectors are over non NA positions
where if lenient is FALSE (default), genotype in genome and rnaseq library
pileup must match exactly for a distance score of 0. Note that A/T == T/A.
If lenient == TRUE, then if the nucleotide is contained in the genotype,
then the distance = 0. For example, A == A, A/T, T/A, A/G, G/A, A/C, C/A.
}
