% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filename_id_subject_mislabel_map}
\alias{filename_id_subject_mislabel_map}
\title{a dataframe to describe the samples to identify}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 41 rows and 4 columns.
}
\usage{
filename_id_subject_mislabel_map
}
\description{
a dataframe to describe the samples to identify
}
\keyword{datasets}
