% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filter_comparison_plt}
\alias{filter_comparison_plt}
\title{plot comparing number of uncertains by position}
\format{
An object of class \code{gg} (inherits from \code{ggplot}) of length 9.
}
\usage{
filter_comparison_plt
}
\description{
plot comparing number of uncertains by position
}
\keyword{datasets}
