% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_genotype_vectors.R
\name{genome_vectors}
\alias{genome_vectors}
\title{generate Snp vectors from each genome}
\usage{
genome_vectors(db_path, tablename)
}
\arguments{
\item{db_path}{a path, eg /path/to/sample_geno_db.sqlite, to where you'd like
to create the sqlite database file}

\item{tablename}{the table from which to extract the SNPS}
}
\value{
a list, length number of individuals in genome database, of snp
vectors
}
\description{
generate Snp vectors from each genome
}
