% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{high_expr_genes}
\alias{high_expr_genes}
\title{A GRangeList of highly expressed genes}
\format{
An object of class \code{GRanges} of length 3688.
}
\usage{
high_expr_genes
}
\description{
A GRangeList of highly expressed genes
}
\keyword{datasets}
