% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_genotype_db_functions.R
\name{highly_expressed_gene_virtual_table}
\alias{highly_expressed_gene_virtual_table}
\title{create virtual table of only highly expressed gene regions in the db}
\usage{
highly_expressed_gene_virtual_table(
  highly_expressed_gene_ranges,
  db_path,
  sample_viewname,
  tablename = "sample_genotype"
)
}
\arguments{
\item{highly_expressed_gene_ranges}{a GRangeList of regions over which to
create a virtual table (a subset of the entire table)}

\item{db_path}{a path, eg /path/to/sample_geno_db.sqlite, to where you'd like
to create the sqlite database file}
}
\description{
create virtual table of only highly expressed gene regions in the db
}
