% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapqOver10_no_secondary_results}
\alias{mapqOver10_no_secondary_results}
\title{results of the most stringent filter on the pileup}
\format{
An object of class \code{list} of length 4.
}
\usage{
mapqOver10_no_secondary_results
}
\description{
results of the most stringent filter on the pileup
}
\keyword{datasets}
