% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAseq_diference_btwn_pileup_df}
\alias{RNAseq_diference_btwn_pileup_df}
\title{comparing RNAseq vectors to quantify similarity/difference btwn snp locations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.
}
\usage{
RNAseq_diference_btwn_pileup_df
}
\description{
comparing RNAseq vectors to quantify similarity/difference btwn snp locations
}
\keyword{datasets}
