% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_genotype_vectors.R
\name{rnaseq_vector}
\alias{rnaseq_vector}
\title{get vector of consensus bases over a given set of SNPs in an RNAseq library}
\usage{
rnaseq_vector(db_path, snp_df, tablename = "rnaseq_pileup")
}
\arguments{
\item{db_path}{path to an RNAseq library pileup sqlite database}

\item{snp_df}{path to the snp dataframe (see function)}

\item{tablename}{the table from which to extract the SNPS}
}
\value{
RNAseq library Snp vector
}
\description{
get vector of consensus bases over a given set of SNPs in an RNAseq library
}
