% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_genotype_vectors.R
\name{sample_snp_loc_from_pileup}
\alias{sample_snp_loc_from_pileup}
\title{extract unique SNP locations from RNAseq library}
\usage{
sample_snp_loc_from_pileup(db_path, tablename = "rnaseq_pileup")
}
\arguments{
\item{db_path}{a path, eg /path/to/sample_geno_db.sqlite, to where you'd like
to create the sqlite database file}

\item{tablename}{name of the rnaseq table. default is rnaseq_pileup}
}
\description{
extract unique SNP locations from RNAseq library
}
