% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_genotype_vectors.R
\name{sample_snp_ranges}
\alias{sample_snp_ranges}
\title{extract unique SNP locations from a sample (view in the sample_genome_db)
as a dataframe}
\usage{
sample_snp_ranges(db_path, tablename)
}
\arguments{
\item{db_path}{a path, eg /path/to/sample_geno_db.sqlite, to where you'd like
to create the sqlite database file}

\item{tablename}{the table from which to extract the SNPS}
}
\description{
extract unique SNP locations from a sample (view in the sample_genome_db)
as a dataframe
}
